Towards a protocol-based implementation.
We try to get rid of interfaces and replace them by protocols much as
possible.  Protocols, which are represented by classes for which
is{instance,subclass} had been overriden and which cannot be
instantiated, are created at the very beginning of startup.  Then, as
the core classes are defined, they are used to extend various protocols.
Only then is RT imported, at which point protocolfns are already
available on most classes.

In the current implementation only ISeq and Seqable were transformed
into protocols (and not even completely, just enough so that REPL
startup is possible).